Increase app memory limit to 256MB
This is enough for my local but seems like it's not enough for
running it on wpc-stage.com. Not sure why this is. But once I profile
it I should be able get a bunch of low-hanging fruit, in terms of
optimizing memory usage, so hopefully not only will it work on staging
but I can actually decrease the setting.